Backport the 1.2 csm_read_secrets changes with a script
Backport the 1.2 changes to the csm_read_secrets module by
providing the same functionality through a python script.
This is being done so as not to interfere with other products
that are using the 1.0 version of csm_read_secrets which is
not backwards compatible with the CSM 1.2 version of this
module.