CASMCMS-8353: Provide authentication environment for cms-meta-tools
In response to CASMTRIAGE-4680 as well as the tightening of permissions
on CASM's artifactory server, cms-meta-tools was upgraded to authenticate
to both DST's artifactory as well as CASM's artifactory.  To authenticate
to CASM's artifactory, we need to set up the environment with
authentication, i.e. user name and password in the Jenkinsfile..

(cherry picked from commit 2f28d48204c254d3e31a66913c2fd52f819fafb7)