Only add SLE specific repos
There are a few packages (such as libcsm) that need to be pulled from the correct, corresponding SLE OS repo. Having other distro's repos added can break package resolution for packages that do care about the distro.

This change removes the guess work, and ensures that just the NOOS repo and the correct corresponding SLE repo is added to the node. The variables used are interpolated by Zypper, and will always match the running distribution.