CASMTRIAGE-5788 Fix YAML Anchor
The `&cfs_image` is breaking the `ncn_sysctl.yml` playbook

```bash
ERROR! Hosts list cannot be empty. Please check your playbook
```

The `&` is causing the YAML to read the value as an anchor, but in this
case it is empty. Wrapping it in quotes prevents YAML from interpolating
the value.

This also fixes the group to include the right hosts ahead of the `&`.