CRAYSAT-1767: Add csm.ncn.sat role to configure SAT
Add a `csm.ncn.sat` role to configure SAT on Kubernetes management NCNs
(masters and workers). This role is being moved from the separate
sat-config-management repository into this repository as part of the
effort to fully merge the SAT product into the CSM product.

The main differences between the Ansible content as it existed in the
SAT product stream and the content added here are as follows:

- The version of the `cray-sat` container image is injected using an
  initContainer in the Helm chart and a loftsman manifest customization
  in the CSM release build process.
- There is no longer a separate package repository in Nexus that must be
  added on management NCNs. The `cray-sat-podman` package is included in
  the CSM noos repository, and it is installed with the existing
  csm.packages role.

Test Description:
Deployed the Helm chart to mug, created a CFS configuration with a layer
that called the `ncn_nodes.yml` playbook, then configured a master,
worker, and storage management NCN image with it. Also configured
ncn-m001 with a session using the same CFS configuration. Verified that
the Ansible was successful and the `/opt/cray/etc/sat/version` file was
written as expected, with the value from `vars/main.yml` for the sat
version taking precedence.