MTL-2348 Fixes QLogic Boots
QLogic cards are not showing in the initramfs, the `qede` module is
failing to load.

This was root caused by the removal of `fastlinq.conf`.

We still have to remove `fastlinq.conf`, as it installs `qedr` and
overrides our blacklist.

The QLogic modules must be specified for the necessary weak-updates to
be installed into the initramfs.

This adds all the modules from `fastlinq.conf`, excluding `qedr`.