CASMTRIAGE-6699 Fix Missing KMP
Kernel modules are not found correctly despite CASMTRIAGE-6617.

CASMTRIAGE-6617 was tested in conjunction with NETETH-2313 and
inadvertently relied on a portion of NETETH-2313's WAR.

This change adds a missing `dracut` call that NETETH-2313 was doing.

After this change, CASMTRIAGE-6617 + CASMTRIAGE-6699 will supersede the
NETETH-2313 WAR - NETETH-2313 is NO LONGER necessary for CSM 1.4.4.